Convert joystick values between -1023 and 1023 to standard -1.0-1.0 values

+++ hri_safety_sense/include/JoystickHandler.h
@@ -38,7 +38,7 @@ namespace hri_safety_sense {
 
 	   private:
 
-		  int32_t getStickValue(JoystickType joystick);
+		  float getStickValue(JoystickType joystick);
 		  int32_t getButtonValue(uint8_t button);
 
 		  rclcpp::Node::SharedPtr rosNode;
